feat(dashboard): move dashboard page to app router
